gh-123361: Split `test_dis.py` into a directory
