+++ Makefile.pre.in
@@ -2438,6 +2438,7 @@ TESTSUBDIRS=	idlelib/idle_test \
 		test/test_ctypes \
 		test/test_dataclasses \
 		test/test_doctest \
+		test/test_dis \
 		test/test_email \
 		test/test_email/data \
 		test/test_free_threading \
